requirements.txt
alembic==1.7.5
certifi==2021.10.8
charset-normalizer==2.0.9
click==7.1.2
docopt==0.6.2
ecdsa==0.17.0
Flask==1.1.4
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jose==1.0.0
Mako==1.1.6
MarkupSafe==2.0.1
pipreqs==0.4.11
psycopg2-binary==2.9.1
pyasn1==0.4.8
python-jose==3.3.0
requests==2.26.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.7
Werkzeug==1.0.1
yarg==0.1.9
zipp==3.7.0
python-dotenv
authlib
six
